@@ -9,13 +9,14 @@
 		"watch": "tsc --watch",
 		"prepublishOnly": "rimraf dist && npm run build && husky install",
 		"lint": "prettier --check '{src,tests}/{**/*,*}.{js,ts,cjs,mjs}' && eslint '{src,tests}/{**/*,*}.{js,ts,cjs,mjs}'",
-		"typecheck": "tsc --noEmit"
+		"typecheck": "tsc --noEmit",
+		"sea": "node buildSea.js"
 	},
 	"bin": {
 		"cross-seed": "dist/cmd.js"
 	},
 	"engines": {
-		"node": ">=20"
+		"node": ">=22"
 	},
 	"files": [
 		"dist/"
@@ -29,7 +30,6 @@
 	"license": "Apache-2.0",
 	"dependencies": {
 		"bencode": "^2.0.1",
-		"better-sqlite3": "^11.5.0",
 		"chalk": "^5.0.0",
 		"commander": "^8.3.0",
 		"fastest-levenshtein": "^1.0.16",
@@ -48,15 +48,17 @@
 		"@types/bencode": "^2.0.1",
 		"@types/lowdb": "^1.0.9",
 		"@types/ms": "^0.7.34",
-		"@types/node": "^20.10.6",
+		"@types/node": "^22.10.2",
 		"@types/xml2js": "^0.4.9",
 		"@types/xmlrpc": "^1.3.6",
 		"@typescript-eslint/eslint-plugin": "^6.16.0",
 		"@typescript-eslint/parser": "^6.16.0",
 		"@vitest/coverage-v8": "^2.0.5",
+		"esbuild": "^0.24.2",
 		"eslint": "^8.56.0",
 		"husky": "^7.0.0",
 		"lint-staged": "^12.1.4",
+		"postject": "^1.0.0-alpha.6",
 		"prettier": "^3.2.5",
 		"rimraf": "^3.0.2",
 		"typescript": "^5.6.2",
